OMEGA-64 Era 960-970: Tensor Web Physics + Φ-Message Lifecycle
- Implemented full tick_physics() with Kuramoto coupling, metabolic decay,
  phase drift, resonance replenish, and compost bridge auto-publishing.
- Added Q10 sine LUT (256-element) and sin_q10() to omega_v2/src/math.rs.
- Added KURAMOTO_COUPLING_BASE, ENERGY_DECAY_PER_TICK, PHASE_DRIFT_Q_SHIFT
  constants to omega_v2/src/constants.rs.
- Fixed dirty flags clearing to happen before early return (reactive reconciliation).
- Expanded lattice tests from 8 to 12: energy_decay, phase_drift, kuramoto_coupling,
  dirty_flags_cleared.
- Updated documentation: COMPLETED_STAGES.md (Era 960-970), PHI_MANIFEST.md.
- Fixed TS declarations: PhaseAgentMinimal size 32 bytes (was 24).

All tests pass: 34/34 omega_v2 + 6/6 omega_core = 40/40.
Clippy clean. WASM build clean.

+++ omega_core/src/utils.rs
@@ -3,7 +3,7 @@ use crate::fixed_point::*;
 
 #[inline(always)]
 pub(crate) fn wrap_phase(value: i16) -> u8 {
-    crate::constants::wrap_index(value as i32, PHASE_LUT_SIZE as i32) as u8
+    crate::constants::wrap_index(value as i32, PHASE_LUT_SIZE) as u8
 }
 
 #[inline(always)]
